refactor Display interface
